Multi-Agents Stock Analysis/requirements.txt
crewai
crewai-tools
python-dotenv
yfinance
litellm